Data Aggregation 1.0
To aggregating data using the NHL API, we can use this code written in Python! Note that the API files are in JSON format and we should finally deal with the JSON files. Also, to download the data, based on the pattern explianed with the API instruction, we can download them in two parts of regular season games and playoffs!