Data Aggregation 1.1 (updated version)
To aggregating data using the NHL API, we can use this code written in Python! Note that the API files are in JSON format and we should finally deal with the JSON files. We can use this code to download all the data (data of regular season games and playoffs) using the API at once in one directory/folder!